+++ packages/plugins/Gfi/src/store/actions/index.ts
@@ -1,5 +1,5 @@
 import debounce from 'lodash.debounce'
-import { Style, Fill, Stroke } from 'ol/style'
+import { Fill, Stroke, Style } from 'ol/style'
 import Overlay from 'ol/Overlay'
 import { GeoJSON } from 'ol/format'
 import { Feature } from 'ol'
@@ -7,12 +7,11 @@ import { Feature as GeoJsonFeature, GeoJsonProperties } from 'geojson'
 import { PolarActionTree } from '@polar/lib-custom-types'
 import getCluster from '@polar/lib-get-cluster'
 import { getTooltip, Tooltip } from '@polar/lib-tooltip'
-import { DragBox } from 'ol/interaction'
-import { platformModifierKeyOnly } from 'ol/events/condition'
-import { getFeatureDisplayLayer, clear } from '../../utils/displayFeatureLayer'
+import { clear, getFeatureDisplayLayer } from '../../utils/displayFeatureLayer'
 import { GfiGetters, GfiState } from '../../types'
 import { getOriginalFeature } from '../../utils/getOriginalFeature'
 import { debouncedGfiRequest } from './debouncedGfiRequest'
+import { setupMultiSelection } from './setupMultiSelection'
 
 // OK for module action set creation
 // eslint-disable-next-line max-lines-per-function
@@ -80,29 +79,7 @@ export const makeActions = () => {
         )
       }
     },
-    setupMultiSelection({ dispatch, getters, rootGetters }) {
-      if (getters.gfiConfiguration.boxSelect) {
-        const dragBox = new DragBox({ condition: platformModifierKeyOnly })
-        dragBox.on('boxend', () =>
-          dispatch('getFeatureInfo', {
-            coordinateOrExtent: dragBox.getGeometry().getExtent(),
-            modifierPressed: true,
-          })
-        )
-        rootGetters.map.addInteraction(dragBox)
-      }
-      if (getters.gfiConfiguration.directSelect) {
-        rootGetters.map.on('click', ({ coordinate, originalEvent }) =>
-          dispatch('getFeatureInfo', {
-            coordinateOrExtent: coordinate,
-            modifierPressed:
-              navigator.userAgent.indexOf('Mac') !== -1
-                ? originalEvent.metaKey
-                : originalEvent.ctrlKey,
-          })
-        )
-      }
-    },
+    setupMultiSelection,
     setupZoomListeners({ dispatch, getters, rootGetters }) {
       if (getters.gfiConfiguration.featureList) {
         this.watch(