+++ packages/plugins/Pins/src/store/index.ts
@@ -53,8 +53,8 @@ export const makeStoreModule = () => {
         rootGetters.map.on('singleclick', async ({ coordinate }) => {
           const isDrawing = interactions.getArray().some(
             (interaction) =>
-              // these indicate other interactions are expected now
-              interaction instanceof Draw ||
+              // @ts-expect-error | internal hack to detect it from @polar/plugin-gfi
+              (interaction instanceof Draw && interaction._isMultiSelect) ||
               interaction instanceof Modify ||
               // @ts-expect-error | internal hack to detect it from Draw plugin
               interaction._isDeleteSelect